Split operator.py into multiple files
