+++ environment.yml
@@ -36,6 +36,7 @@ dependencies:
     # For testing
     - packaging
     - pytest-cov
+    - tomli
     # For debugging
     - icecream
     - ipykernel
