+++ graphblas/agg/__init__.py
@@ -111,6 +111,6 @@ def __getattr__(key):
     raise AttributeError(f"module {__name__!r} has no attribute {key!r}")
 
 
-from ..core import agg  # noqa: E402 isort:skip
+from ..core import operator  # noqa: E402 isort:skip
 
-del agg
+del operator
