+++ graphblas/tests/test_op.py
@@ -1332,6 +1332,8 @@ def test_deprecated():
         gb.op.secondj
     with pytest.warns(DeprecationWarning, match="please use"):
         gb.agg.argmin
+    with pytest.warns(DeprecationWarning, match="please use"):
+        import graphblas.core.agg  # noqa: F401
 
 
 def test_is_idempotent():
