+++ scripts/create_pickle.py
@@ -6,7 +6,7 @@ Python version is used to create them.
 """
 import argparse
 import pickle
-from pathlib import PurePath
+from pathlib import Path
 
 import graphblas as gb
 from graphblas.tests.test_pickle import *
@@ -158,7 +158,7 @@ if __name__ == "__main__":
         extra = "-vanilla"
     else:
         extra = ""
-    path = PurePath(gb.tests.__file__).parent
+    path = Path(gb.tests.__file__).parent
     pickle1(path / f"pickle1{extra}.pkl")
     pickle2(path / f"pickle2{extra}.pkl")
     pickle3(path / f"pickle3{extra}.pkl")
